Feature/AI_Chatbot/requirement.txt
# Core AWS and SageMaker dependencies
boto3==1.34.144
botocore==1.34.144
sagemaker==2.231.0

# Data processing and analysis
pandas==2.0.3
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.1

# Vector databases and search
pinecone==5.3.1

# Cloud storage and databases
firebase-admin==6.0.1
google-cloud-storage==2.18.2

# Machine learning utilities
mlxtend==0.23.1

# AI and LLM integrations
openai==1.50.2
google-generativeai==0.8.3
google-genai==1.0.0
transformers==4.44.2
torch==2.3.1

# Async and networking
aiohttp==3.10.5
asyncio==3.4.3

# Logging and monitoring
structlog==24.4.0
prometheus-client==0.20.0

# Testing and development
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
black==24.8.0
flake8==7.1.1

# Security and validation
pydantic==2.9.2
cryptography==43.0.1

# Utilities
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
